Changes for reporting, softVerify and verify

+++ cypress.config.js
@@ -1,60 +1,102 @@
-const { defineConfig } = require("cypress");
 const {
-  addCucumberPreprocessorPlugin, afterRunHandler
+  defineConfig
+} = require("cypress");
+const {
+  addCucumberPreprocessorPlugin,
+  afterRunHandler,
 } = require("@badeball/cypress-cucumber-preprocessor");
 const {
   preprocessor,
 } = require("@badeball/cypress-cucumber-preprocessor/browserify");
+const {
+  rmdir
+} = require("fs");
+const mysql = require("mysql");
+const fs = require("fs");
+const path = require("path");
+const allureWriter = require("@shelex/cypress-allure-plugin/writer");
 const report = require("multiple-cucumber-html-reporter");
+const base64 = require("js-base64");
 let browserDetails;
-const fs = require('fs')
-const path = require('path')
-const base64 = require('js-base64')
-
 
 async function setupNodeEvents(on, config) {
-  // This is required for the preprocessor to be able to generate JSON reports after each run, and more,
-  await addCucumberPreprocessorPlugin(on, config,{
-    omitAfterRunHandler: true
+  await addCucumberPreprocessorPlugin(on, config, {
+    omitAfterRunHandler: true,
   });
-
   on("file:preprocessor", preprocessor(config));
-
+  on("task", {
+    deleteDownloads() {
+      return new Promise((resolve) => {
+        rmdir("cypress/downloads", {
+          recursive: true
+        }, () => {
+          resolve(null);
+        });
+      });
+    },
+  });
+  on("task", {
+    queryDb: (query) => {
+      const connection = mysql.createConnection(config.env.db);
+      connection.connect();
+      return new Promise((resolve, reject) => {
+        connection.query(query, (error, results) => {
+          if (error) reject(error);
+          else {
+            connection.end();
+            return resolve(results);
+          }
+        });
+      });
+    },
+  });
+  allureWriter(on, config);
   on("after:run", async (results) => {
     await afterRunHandler(config);
-    browserDetails = results
-    renameCucumberJSONFile('cypress/test-outputs/json-reports',results.endedTestsAt)
-    generateHTMLReport()
-  })
-
+    browserDetails = results;
+    renameCucumberJSONFile("cypress/test-output/json-reports", results.endedTestsAt);
+    generateHTMLReport();
+  });
   return config;
 }
 
 module.exports = defineConfig({
-  defaultCommandTimeout:8000,
-  requestTimeout:20000,
-  responseTimeout:60000,
-  viewportHeight:1080,
-  viewportWidth:1080,
-  video:false,
-  videoCompression:40,
-  trashAssetsBeforeRuns:false,
+  pageLoadTimeout: 360000,
+  defaultCommandTimeout: 8000,
+  requestTimeout: 20000,
+  responseTimeout: 60000,
+  viewportHeight: 1080,
+  viewportWidth: 1920,
+  video: false,
+  videoCompression: 40,
+  trashAssetsBeforeRuns: false,
+  experimentalMemoryManagement: true,
+  projectId: "cypress-test-automation",
   e2e: {
-    testIsolation:false,
-    inlineAssets:true,
-    screenshotsFolder:'cypress/test-outputs/screenshots',
-    screenshotOnRunFailure:true,
-    setupNodeEvents,
-    specPattern: "cypress/integration/features/*.feature",
     baseUrl: "https://www.saucedemo.com",
     chromeWebSecurity: false,
+    specPattern: "cypress/e2e/features/*.feature",
+    testIsolation: false,
+    inlineAssets: true,
+    screenshotsFolder: "cypress/test-output/screenshots",
+    screenshotOnRunFailure: true,
+    env: {
+      softAssertionCounter: 1,
+      allureReuseAfterSpec: true,
+      allureLogCypress: true,
+      allureResultsPath: "cypress/test-output/allure-results",
+      allureAttachRequests: true,
+    },
+    setupNodeEvents,
   },
 });
 
+//This method generates HTML Report after the Execution in "test-output/html-report" folder
 function generateHTMLReport() {
-  const cucumberJsonDir = './cypress/test-outputs/json-reports'
-  const htmlReportDir = './cypress/test-outputs/html-report'
-  const osName = browserDetails.osName=="darwin"?"osx":browserDetails.osName
+  const cucumberJsonDir = "./cypress/test-output/json-reports";
+  const htmlReportDir = "./cypress/test-output/html-report";
+  const osName = browserDetails.osName == "darwin" ? "osx" : browserDetails.osName;
+  processCucumberJsonFiles(cucumberJsonDir);
   report.generate({
     openReportInBrowser: true,
     jsonDir: cucumberJsonDir,
@@ -63,8 +105,8 @@ function generateHTMLReport() {
     displayDuration: true,
     displayReportTime: true,
     removeFolders: true,
-    pageTitle: 'Execution Report',
-    reportName: `Execution Report - ${new Date().toLocaleString()}`,
+    pageTitle: "System-Test Report",
+    reportName: `System - Test Report - ${new Date().toLocaleString()}`,
     metadata: {
       browser: {
         name: browserDetails.browserName,
@@ -73,23 +115,70 @@ function generateHTMLReport() {
       device: "Local test machine",
       platform: {
         name: osName,
-        version: browserDetails.osVersion
+        version: browserDetails.osVersion,
       },
     },
-  })
+  });
 }
 
-function renameCucumberJSONFile(folderPath,timestamp){
-  const newFileName = path.join(folderPath,`cucumber-report-${timestamp}.json`);
-  console.log('file path : ',path.join(folderPath,'cucumberReport.json'))
+function sanitizeTimestamp(timestamp) {
+  // Replace invalid characters for filenames
+  return timestamp.replace(/[:]/g, "-");
+}
+
+function renameCucumberJSONFile(folderPath, timestamp) {
+  const sanitizedTimestamp = sanitizeTimestamp(timestamp);
+  const originalFilePath = path.join(folderPath, "cucumber.json");
+  const newFileName = path.join(folderPath, `report-${sanitizedTimestamp}.json`);
+  console.log(`Original File Path: ${originalFilePath}`);
+  console.log(`New File Path: ${newFileName}`);
   try {
-    const cucumberData = fs.readFileSync(path.join(folderPath,'cucumberReport.json'),'utf-8');
-    console.log("cucumber data: ",cucumberData)
-    fs.writeFileSync(newFileName,cucumberData);
-    console.log(`Created report: ${newFileName}`)
-    fs.unlinkSync(path.join(folderPath,'cucumberReport.json'));
-    console.log('Deleted cucumberReport.json');
+    if (!fs.existsSync(originalFilePath)) {
+      console.error("Error: cucumber.json file does not exist");
+      return;
+    }
+    const cucumberData = fs.readFileSync(originalFilePath, "utf-8");
+    fs.writeFileSync(newFileName, cucumberData);
+    console.log(`Created report: ${newFileName}`);
+    try {
+      fs.unlinkSync(originalFilePath);
+      console.log("Deleted cucumber.json");
+    } catch (error) {
+      console.error("Error deleting cucumber.json:", error);
+    }
   } catch (error) {
-    console.error('Error deleting cucumberReport.json: ',error)
+    console.error("Error during file operations:", error);
   }
+}
+
+
+function decodeBase64(data) {
+  return base64.decode(data);
+}
+
+function processCucumberJsonFiles(folderPath) {
+  const files = fs.readdirSync(folderPath);
+  files.forEach((fileName) => {
+    if (path.extname(fileName) === ".json") {
+      const filePath = path.join(folderPath, fileName);
+      const fileContents = fs.readFileSync(filePath, "utf-8");
+      const cucumberReport = JSON.parse(fileContents);
+      cucumberReport.forEach((feature) => {
+        feature.elements.forEach((scenario) => {
+          scenario.steps.forEach((step) => {
+            if (step.embeddings) {
+              step.embeddings.forEach((embedding) => {
+                const decodedData = decodeBase64(embedding.data);
+                if (decodedData == "Soft Assertion failure") {
+                  step.result.status = "failed";
+                  step.result.error_message = "Soft assertion error";
+                }
+              });
+            }
+          });
+        });
+      });
+      fs.writeFileSync(filePath, JSON.stringify(cucumberReport, null, 2));
+    }
+  });
 }
